requirements.txt
kiteconnect==5.0.1
pandas==2.2.3
numpy==2.1.3
pyyaml==6.0.2
rich==13.9.4
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3
websocket-client==1.8.0
